Create an automated bootstrap process
Allows a new user to run a single command that automates most of the
instructions/manual work in this section of the README:

https://github.com/mozilla/payments-env#development

Most of these were commands anyway, but I put them in a Python script
with prompts and handy messages.

It's tested solely on OS X, but it should work on any machine with the required
software (Python/Docker/Node) installed. Windows support isn't guaranteed but
actually should be good, except I don't know how to set up environment variables
on Windows.

Closes #54.
Fixes #50, as this adds in an env file during bootstrap and tells the user
how to source it.